in titlebar(), space should be an int instead of a size_t
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@2162 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/winio.c
@@ -2224,7 +2224,7 @@ void statusq_abort(void)
 
 void titlebar(const char *path)
 {
-    size_t space;
+    int space;
 	/* The space we have available for display. */
     size_t verlen = strlen(VERMSG) + 1;
 	/* The length of the version message. */
